break out Custom Field Names tab settings
