+++ src/settings.js
@@ -2,6 +2,7 @@ import { PluginSettingTab, Setting, FuzzySuggestModal, MarkdownView } from 'obsi
 import { createExampleSettings } from './settingsTabExample';
 import { createAPISettings } from './settingsTabAPISettings';
 import { createFolderSettings } from './settingsTabFolderImages';
+import { createCustomFieldsSettings } from './settingsTabCustomFieldNames';
 
 const DEFAULT_SETTINGS = {
     apiProvider: 'all',
@@ -89,43 +90,6 @@ class FolderSuggestModal extends FuzzySuggestModal {
     }
 }
 
-// Helper functions
-function arrayToString(arr) {
-    return Array.isArray(arr) ? arr.join(', ') : arr;
-}
-
-function stringToArray(str) {
-    return str.split(',')
-        .map(s => s.trim())
-        .filter(s => s.length > 0);
-}
-
-function validateFieldNames(settings, allFields, currentField, newNames) {
-    // Check for valid characters in field names (alphanumeric, dashes, underscores only)
-    const validNamePattern = /^[a-zA-Z0-9_-]+$/;
-    const invalidNames = newNames.filter(name => !validNamePattern.test(name));
-    if (invalidNames.length > 0) {
-        return {
-            isValid: false,
-            message: `Invalid characters in field names (only letters, numbers, dashes, and underscores allowed): ${invalidNames.join(', ')}`
-        };
-    }
-
-    // Then check for duplicates
-    const otherFields = allFields.filter(f => f !== currentField);
-    const otherFieldNames = otherFields.flatMap(f => settings[f]);
-    const duplicates = newNames.filter(name => otherFieldNames.includes(name));
-    
-    if (duplicates.length > 0) {
-        return {
-            isValid: false,
-            message: `Duplicate field names found: ${duplicates.join(', ')}`
-        };
-    }
-    
-    return { isValid: true };
-}
-
 function migrateSettings(settings) {
     const fieldsToMigrate = [
         'customBannerField',
@@ -143,24 +107,6 @@ function migrateSettings(settings) {
     });
 
     return settings;
-}
-
-class PixelBannerPlugin extends Plugin {
-    async loadSettings() {
-        this.settings = Object.assign({}, DEFAULT_SETTINGS, await this.loadData());
-        this.settings = migrateSettings(this.settings);
-        
-        if (!Array.isArray(this.settings.folderImages)) {
-            this.settings.folderImages = [];
-        }
-
-        if (this.settings.folderImages) {
-            this.settings.folderImages.forEach(folderImage => {
-                folderImage.imageDisplay = folderImage.imageDisplay || 'cover';
-                folderImage.imageRepeat = folderImage.imageRepeat || false;
-            });
-        }
-    }
 }
 
 class PixelBannerSettingTab extends PluginSettingTab {
@@ -191,7 +137,7 @@ class PixelBannerSettingTab extends PluginSettingTab {
 
         // Custom Fields tab content
         const customFieldsTab = tabContentContainer.createEl('div', { cls: 'tab-content', attr: { 'data-tab': 'Custom Field Names' } });
-        this.createCustomFieldsSettings(customFieldsTab);
+        createCustomFieldsSettings(customFieldsTab, this.plugin);
 
         // API Settings tab content
         const apiTab = tabContentContainer.createEl('div', { cls: 'tab-content', attr: { 'data-tab': 'API Settings' } });
@@ -345,7 +291,6 @@ class PixelBannerSettingTab extends PluginSettingTab {
                 .onClick(async () => {
                     this.plugin.settings.imageDisplay = DEFAULT_SETTINGS.imageDisplay;
                     await this.plugin.saveSettings();
-                    this.plugin.updateAllBanners();
                     const dropdownEl = button.extraSettingsEl.parentElement.querySelector('select');
                     dropdownEl.value = DEFAULT_SETTINGS.imageDisplay;
                     dropdownEl.dispatchEvent(new Event('change'));
@@ -1041,210 +986,6 @@ class PixelBannerSettingTab extends PluginSettingTab {
                         toggleComponent.setValue(DEFAULT_SETTINGS.showReleaseNotes);
                     }
                 }));
-    }
-
-    createCustomFieldsSettings(containerEl) {
-        // section callout
-        const calloutEl = containerEl.createEl('div', { cls: 'tab-callout' });
-        calloutEl.createEl('div', { text: 'Customize the frontmatter field names used for the banner and Y-position. You can define multiple names for each field, separated by commas. Field names can only contain letters, numbers, dashes, and underscores. Example: "banner, pixel-banner, header_image" could all be used as the banner field name.' });
-
-        const customFields = [
-            {
-                setting: 'customBannerField',
-                name: 'Banner Field Names',
-                desc: 'Set custom field names for the banner in frontmatter (comma-separated)',
-                values: '[[image.png]], "images/image.jpg"',
-                placeholder: 'banner, pixel-banner, header-image'
-            },
-            {
-                setting: 'customYPositionField',
-                name: 'Y-Position Field Names',
-                desc: 'Set custom field names for the Y-position in frontmatter (comma-separated)',
-                values: '5, 70, 100',
-                placeholder: 'banner-y, y-position, banner-offset'
-            },
-            {
-                setting: 'customXPositionField',
-                name: 'X-Position Field Names',
-                desc: 'Set custom field names for the X-position in frontmatter (comma-separated)',
-                values: '0, 30, 90',
-                placeholder: 'banner-x, x-position, banner-offset-x'
-            },
-            {
-                setting: 'customContentStartField',
-                name: 'Content Start Position Field Names',
-                desc: 'Set custom field names for the content start position in frontmatter (comma-separated)',
-                values: '75, 150, 450',
-                placeholder: 'content-start, start-position, content-offset'
-            },
-            {
-                setting: 'customImageDisplayField',
-                name: 'Image Display Field Names',
-                desc: 'Set custom field names for the image display in frontmatter (comma-separated)',
-                values: 'cover, contain, auto, 200%, 70%',
-                placeholder: 'banner-display, image-display, display-mode'
-            },
-            {
-                setting: 'customImageRepeatField',
-                name: 'Image Repeat Field Names',
-                desc: 'Set custom field names for the image repeat in frontmatter (comma-separated)',
-                values: 'true, false',
-                placeholder: 'banner-repeat, image-repeat, repeat-image'
-            },
-            {
-                setting: 'customBannerHeightField',
-                name: 'Banner Height Field Names',
-                desc: 'Set custom field names for the banner height in frontmatter (comma-separated)',
-                values: '100, 300, 700',
-                placeholder: 'banner-height, image-height, header-height'
-            },
-            {
-                setting: 'customFadeField',
-                name: 'Fade Field Names',
-                desc: 'Set custom field names for the fade effect in frontmatter (comma-separated)',
-                values: '-1000, -100, 100',
-                placeholder: 'banner-fade, fade-effect, image-fade'
-            },
-            {
-                setting: 'customBorderRadiusField',
-                name: 'Border Radius Field Names',
-                desc: 'Set custom field names for the border radius in frontmatter (comma-separated)',
-                values: '0, 17, 30, 50',
-                placeholder: 'banner-radius, border-radius, banner-corner-radius'
-            },
-            {
-                setting: 'customTitleColorField',
-                name: 'Inline Title Color Field Names',
-                desc: 'Set custom field names for the inline title color in frontmatter (comma-separated)',
-                values: 'red, papayawhip, "#7f6df2", "#ffa500"',
-                placeholder: 'banner-title-color, title-color, header-color'
-            },
-            {
-                setting: 'customBannerShuffleField',
-                name: 'Banner Shuffle Field Names',
-                desc: 'Set custom field names for the banner shuffle in frontmatter (comma-separated)',
-                values: '"pixel-banner-images", "images/llamas"',
-                placeholder: 'banner-shuffle, shuffle-folder, random-image-folder'
-            },
-            {
-                setting: 'customBannerIconField',
-                name: 'Banner Icon Field Names',
-                desc: 'Set custom field names for the banner icon in frontmatter (comma-separated)',
-                values: '🤣, 💥, ✨',
-                placeholder: 'banner-icon, pixel-icon, header-icon'
-            },
-            {
-                setting: 'customBannerIconSizeField',
-                name: 'Banner Icon Size Field Names',
-                desc: 'Set custom field names for the banner icon size in frontmatter (comma-separated)',
-                values: '70, 100, 150',
-                placeholder: 'banner-icon-size, icon-size, header-icon-size'
-            },
-            {
-                setting: 'customBannerIconXPositionField',
-                name: 'Banner Icon X Position Field Names',
-                desc: 'Set custom field names for the banner icon x position in frontmatter (comma-separated)',
-                values: '25, 50, 75 (value between 0 and 100)',
-                placeholder: 'banner-icon-x, icon-x, header-icon-x'
-            },
-            {
-                setting: 'customBannerIconOpacityField',
-                name: 'Banner Icon Opacity Field Names',
-                desc: 'Set custom field names for the banner icon opacity in frontmatter (comma-separated)',
-                values: '100, 75, 50 (value between 0 and 100)',
-                placeholder: 'banner-icon-opacity, icon-opacity, header-icon-opacity'
-            },
-            {
-                setting: 'customBannerIconColorField',
-                name: 'Banner Icon Text Color Field Names',
-                desc: 'Set custom field names for the banner icon text color in frontmatter (comma-separated)',
-                values: 'white, papayawhip, "#7f6df2", "#ffa500"',
-                placeholder: 'banner-icon-color, icon-color, header-icon-color'
-            },
-            {
-                setting: 'customBannerIconBackgroundColorField',
-                name: 'Banner Icon Background Color Field Names',
-                desc: 'Set custom field names for the banner icon background color in frontmatter (comma-separated)',
-                values: 'transparent, papayawhip, "#7f6df2", "#ffa500"',
-                placeholder: 'banner-icon-background-color, icon-background-color, header-icon-background-color'
-            },
-            {
-                setting: 'customBannerIconPaddingField',
-                name: 'Banner Icon Padding Field Names',
-                desc: 'Set custom field names for the banner icon padding in frontmatter (comma-separated)',
-                values: '0, 10, 20',
-                placeholder: 'banner-icon-padding, icon-padding, header-icon-padding'
-            },
-            {
-                setting: 'customBannerIconBorderRadiusField',
-                name: 'Banner Icon Border Radius Field Names',
-                desc: 'Set custom field names for the banner icon border radius in frontmatter (comma-separated)',
-                values: '0, 17, 30, 50',
-                placeholder: 'banner-icon-border-radius, icon-border-radius, header-icon-border-radius'
-            },
-            {
-                setting: 'customBannerIconVeritalOffsetField',
-                name: 'Banner Icon Vertical Offset Field Names',
-                desc: 'Set custom field names for the banner icon vertical offset in frontmatter (comma-separated)',
-                values: '0, 10, 20',
-                placeholder: 'banner-icon-vertical-offset, icon-vertical-offset, header-icon-vertical-offset'
-            },
-        ];
-
-        customFields.forEach(field => {
-            const settingContainer = new Setting(containerEl)
-                .setName(field.name)
-                .setDesc(field.desc);
-
-            // Add example values if they exist
-            if (field.values) {
-                settingContainer.descEl.createEl('div', {
-                    text: `example frontmatter values: ${field.values}`,
-                    cls: 'setting-item-description pixel-banner-example-values'
-                });
-            }
-
-            settingContainer
-                .addText(text => {
-                    text
-                        .setPlaceholder(field.placeholder)
-                        .setValue(arrayToString(this.plugin.settings[field.setting]))
-                        .onChange(async (value) => {
-                            const newNames = stringToArray(value);
-                            const validation = validateFieldNames(
-                                this.plugin.settings,
-                                customFields.map(f => f.setting),
-                                field.setting,
-                                newNames
-                            );
-
-                            if (validation.isValid) {
-                                this.plugin.settings[field.setting] = newNames;
-                                await this.plugin.saveSettings();
-                            } else {
-                                new Notice(validation.message);
-                                text.setValue(arrayToString(this.plugin.settings[field.setting]));
-                            }
-                        });
-                    text.inputEl.style.width = '220px';
-                })
-                .addExtraButton(button => button
-                    .setIcon('reset')
-                    .setTooltip('Reset to default')
-                    .onClick(async () => {
-                        this.plugin.settings[field.setting] = DEFAULT_SETTINGS[field.setting];
-                        await this.plugin.saveSettings();
-                        
-                        // Update only this specific setting
-                        const settingEl = button.extraSettingsEl.parentElement;
-                        const textInput = settingEl.querySelector('input[type="text"]');
-                        textInput.value = arrayToString(DEFAULT_SETTINGS[field.setting]);
-                        
-                        // Trigger the change event to update the plugin's state
-                        const event = new Event('input', { bubbles: true, cancelable: true });
-                        textInput.dispatchEvent(event);
-                    }));
-        });
     }
 }
